Some corrections on hyperspace balancing which got lost in the merge.

+++ src/Pi.h
@@ -97,7 +97,7 @@ public:
 	}
 	static void SetMouseGrab(bool on);
 	static void BoinkNoise();
-	static float CalcHyperspaceRangeMax(int hyperclass, int total_mass_in_tonnes);
+	static float CalcHyperspaceRangeMax(int hyperclass, float total_mass_in_tonnes);
 	static float CalcHyperspaceRange(int hyperclass, float total_mass_in_tonnes, int fuel);
 	static float CalcHyperspaceDuration(int hyperclass, int total_mass_in_tonnes, float dist);
 	static float CalcHyperspaceFuelOut(int hyperclass, float dist, float hyperspace_range_max);
