+++ src/Ship.cpp
@@ -380,7 +380,7 @@ vector3d Ship::GetMaxThrust(const vector3d &dir) const
 double Ship::GetAccelMin() const
 {
 	const ShipType &stype = GetShipType();
-	float val = fabs(stype.linThrust[ShipType::THRUSTER_UP]);
+	double val = fabs(stype.linThrust[ShipType::THRUSTER_UP]);
 	val = std::min(val, fabs(stype.linThrust[ShipType::THRUSTER_RIGHT]));
 	val = std::min(val, fabs(stype.linThrust[ShipType::THRUSTER_LEFT]));
 	return val / GetMass();
